requirements.txt
django
django-allauth
python-ldap